% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dms_to_dd.R
\name{convert_dms_to_dd}
\alias{convert_dms_to_dd}
\title{Convert Degrees Minutes Seconds (DMS) to Decimal Degrees (DD)}
\usage{
convert_dms_to_dd(degrees, minutes, seconds)
}
\arguments{
\item{degrees}{The degree value from DMS data}

\item{minutes}{The minutes value from DMS data}

\item{seconds}{The seconds value from DMS data}
}
\value{
The corresponding DD value
}
\description{
Convert Degrees Minutes Seconds (DMS) to Decimal Degrees (DD)
}
\examples{
convert_dms_to_dd(11, 34, 19)
}
