% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_degrees.R
\name{extract_degrees}
\alias{extract_degrees}
\title{Extract degrees from latitude / longitude string}
\usage{
extract_degrees(dms_string)
}
\arguments{
\item{dms_string}{The DMS string}
}
\value{
The numeric degrees value
}
\description{
Extract degrees from latitude / longitude string
}
\examples{
extract_degrees("11° 34 ' 19\"")
}
