% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_minutes.R
\name{extract_minutes}
\alias{extract_minutes}
\title{Extract minutes from latitude / longitude string}
\usage{
extract_minutes(dms_string)
}
\arguments{
\item{dms_string}{The DMS string}
}
\value{
The numeric minutes value
}
\description{
Extract minutes from latitude / longitude string
}
\examples{
extract_minutes("11° 34 ' 19\"")
}
