% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seconds.R
\name{extract_seconds}
\alias{extract_seconds}
\title{Extract seconds from latitude / longitude string}
\usage{
extract_seconds(dms_string)
}
\arguments{
\item{dms_string}{The DMS string}
}
\value{
The numeric seconds value
}
\description{
Extract seconds from latitude / longitude string
}
\examples{
extract_seconds("11° 34 \' 19\"")
}
